--- a/src/main/java/org/apache/sysds/runtime/io/WriterBinaryBlock.java
+++ b/src/main/java/org/apache/sysds/runtime/io/WriterBinaryBlock.java
@@ -37,27 +37,35 @@ import org.apache.sysds.runtime.util.HDFSTool;
 public class WriterBinaryBlock extends MatrixWriter {
 	protected int _replication = -1;
 
+	protected static int jobUse = 0;
+	protected static JobConf job = new JobConf(ConfigurationManager.getCachedJobConf());
+
 	public WriterBinaryBlock(int replication) {
 		_replication = replication;
+
+		jobUse ++;
+		if(jobUse > 15){
+			// job =  new JobConf();
+			job = new JobConf(ConfigurationManager.getCachedJobConf());
+			jobUse = 0;
+		}
 	}
 
 	@Override
 	public final void writeMatrixToHDFS(MatrixBlock src, String fname, long rlen, long clen, int blen, long nnz,
 		boolean diag) throws IOException, DMLRuntimeException {
 		// prepare file access
-		JobConf job = new JobConf(ConfigurationManager.getCachedJobConf());
 		Path path = new Path(fname);
-		FileSystem fs = IOUtilFunctions.getFileSystem(path, job);
 
 		// if the file already exists on HDFS, remove it.
-		HDFSTool.deleteFileIfExistOnHDFS(fname);
+		HDFSTool.deleteFileIfExistOnHDFS(path, job);
 
 		// set up preferred custom serialization framework for binary block format
 		if(HDFSTool.USE_BINARYBLOCK_SERIALIZATION)
 			HDFSTool.addBinaryBlockSerializationFramework(job);
 
 		if(src instanceof CompressedMatrixBlock) {
-			if(ConfigurationManager.getCompilerConfigFlag(ConfigType.PARALLEL_CP_WRITE_BINARYFORMATS)){
+			if(ConfigurationManager.getCompilerConfigFlag(ConfigType.PARALLEL_CP_WRITE_BINARYFORMATS)) {
 				LOG.debug("Multi threaded decompression");
 				// parallel
 				src = CompressedMatrixBlock.getUncompressed(src, "binary write",
@@ -71,17 +79,15 @@ public class WriterBinaryBlock extends MatrixWriter {
 
 		// core write sequential/parallel
 		if(diag)
-			writeDiagBinaryBlockMatrixToHDFS(path, job, fs, src, rlen, clen, blen);
+			writeDiagBinaryBlockMatrixToHDFS(path, job, src, rlen, clen, blen);
 		else
-			writeBinaryBlockMatrixToHDFS(path, job, fs, src, rlen, clen, blen);
+			writeBinaryBlockMatrixToHDFS(path, job, src, rlen, clen, blen);
 
-		IOUtilFunctions.deleteCrcFilesFromLocalFileSystem(fs, path);
 	}
 
 	@Override
 	public final void writeEmptyMatrixToHDFS(String fname, long rlen, long clen, int blen)
 		throws IOException, DMLRuntimeException {
-		JobConf job = new JobConf(ConfigurationManager.getCachedJobConf());
 		Path path = new Path(fname);
 		FileSystem fs = IOUtilFunctions.getFileSystem(path, job);
 		final Writer writer = IOUtilFunctions.getSeqWriter(path, job, _replication);
@@ -98,26 +104,22 @@ public class WriterBinaryBlock extends MatrixWriter {
 		IOUtilFunctions.deleteCrcFilesFromLocalFileSystem(fs, path);
 	}
 
-	protected void writeBinaryBlockMatrixToHDFS(Path path, JobConf job, FileSystem fs, MatrixBlock src, long rlen,
+	protected void writeBinaryBlockMatrixToHDFS(Path path, JobConf job,  MatrixBlock src, long rlen,
 		long clen, int blen) throws IOException, DMLRuntimeException {
 		// sequential write
-		writeBinaryBlockMatrixToSequenceFile(path, job, fs, src, blen, 0, (int) rlen);
+		writeBinaryBlockMatrixToSequenceFile(path, job, src, blen, 0, (int) rlen);
+		IOUtilFunctions.deleteCrcFilesFromLocalFileSystem(job, path);
 	}
 
-	protected final void writeBinaryBlockMatrixToSequenceFile(Path path, JobConf job, FileSystem fs, MatrixBlock src,
+	protected final void writeBinaryBlockMatrixToSequenceFile(Path path, JobConf job,  MatrixBlock src,
 		int blen, int rl, int ru) throws IOException {
 		boolean sparse = src.isInSparseFormat();
-		int rlen = src.getNumRows();
-		int clen = src.getNumColumns();
-
+		final int rlen = src.getNumRows();
+		final int clen = src.getNumColumns();
 		final Writer writer = IOUtilFunctions.getSeqWriter(path, job, _replication);
 
-		try { // 2) bound check for src block
-			if(src.getNumRows() > rlen || src.getNumColumns() > clen) {
-				throw new IOException("Matrix block [1:" + src.getNumRows() + ",1:" + src.getNumColumns() + "] "
-					+ "out of overall matrix range [1:" + rlen + ",1:" + clen + "].");
-			}
-
+		try { 
+	
 			// 3) reblock and write
 			MatrixIndexes indexes = new MatrixIndexes();
 
@@ -133,10 +135,9 @@ public class WriterBinaryBlock extends MatrixWriter {
 
 				// create and write sub-blocks of matrix
 				for(int blockRow = rl / blen; blockRow < (int) Math.ceil(ru / (double) blen); blockRow++) {
-					for(int blockCol = 0; blockCol < (int) Math.ceil(src.getNumColumns() / (double) blen); blockCol++) {
-						int maxRow = (blockRow * blen + blen < src.getNumRows()) ? blen : src.getNumRows() - blockRow * blen;
-						int maxCol = (blockCol * blen + blen < src.getNumColumns()) ? blen : src.getNumColumns() -
-							blockCol * blen;
+					for(int blockCol = 0; blockCol < (int) Math.ceil(clen / (double) blen); blockCol++) {
+						int maxRow = (blockRow * blen + blen < rlen) ? blen : rlen - blockRow * blen;
+						int maxCol = (blockCol * blen + blen < clen) ? blen : clen - blockCol * blen;
 
 						int row_offset = blockRow * blen;
 						int col_offset = blockCol * blen;
@@ -144,7 +145,7 @@ public class WriterBinaryBlock extends MatrixWriter {
 						// get reuse matrix block
 						MatrixBlock block = getMatrixBlockForReuse(blocks, maxRow, maxCol, blen);
 
-						// copy submatrix to block
+						// copy sub matrix to block
 						src.slice(row_offset, row_offset + maxRow - 1, col_offset, col_offset + maxCol - 1, block);
 
 						// append block to sequence file
@@ -162,16 +163,17 @@ public class WriterBinaryBlock extends MatrixWriter {
 		}
 	}
 
-	protected final void writeDiagBinaryBlockMatrixToHDFS(Path path, JobConf job, FileSystem fs, MatrixBlock src,
-		long rlen, long clen, int blen) throws IOException, DMLRuntimeException {
+	protected final void writeDiagBinaryBlockMatrixToHDFS(Path path, JobConf job,  MatrixBlock src,
+		long rlen, long clen, int blen) throws IOException {
 		boolean sparse = src.isInSparseFormat();
-
+		final int nRow = src.getNumRows();
+		final int nCol = src.getNumColumns();
 		final Writer writer = IOUtilFunctions.getSeqWriter(path, job, _replication);
 
 		try {
 			// 2) bound check for src block
-			if(src.getNumRows() > rlen || src.getNumColumns() > clen) {
-				throw new IOException("Matrix block [1:" + src.getNumRows() + ",1:" + src.getNumColumns() + "] "
+			if(nRow > rlen || nCol > clen) {
+				throw new IOException("Matrix block [1:" + nRow + ",1:" + nCol + "] "
 					+ "out of overall matrix range [1:" + rlen + ",1:" + clen + "].");
 			}
 
@@ -188,13 +190,12 @@ public class WriterBinaryBlock extends MatrixWriter {
 				MatrixBlock[] blocks = createMatrixBlocksForReuse(rlen, clen, blen, sparse, src.getNonZeros());
 				MatrixBlock emptyBlock = new MatrixBlock();
 
-				// create and write subblocks of matrix
-				for(int blockRow = 0; blockRow < (int) Math.ceil(src.getNumRows() / (double) blen); blockRow++) {
+				// create and write sub blocks of the matrix
+				for(int blockRow = 0; blockRow < (int) Math.ceil(nRow / (double) blen); blockRow++) {
 
-					for(int blockCol = 0; blockCol < (int) Math.ceil(src.getNumColumns() / (double) blen); blockCol++) {
-						int maxRow = (blockRow * blen + blen < src.getNumRows()) ? blen : src.getNumRows() - blockRow * blen;
-						int maxCol = (blockCol * blen + blen < src.getNumColumns()) ? blen : src.getNumColumns() -
-							blockCol * blen;
+					for(int blockCol = 0; blockCol < (int) Math.ceil(nCol / (double) blen); blockCol++) {
+						int maxRow = (blockRow * blen + blen < nRow) ? blen : nRow - blockRow * blen;
+						int maxCol = (blockCol * blen + blen < nCol) ? blen : nCol - blockCol * blen;
 						MatrixBlock block = null;
 
 						if(blockRow == blockCol) { // block on diagonal
@@ -204,7 +205,7 @@ public class WriterBinaryBlock extends MatrixWriter {
 							// get reuse matrix block
 							block = getMatrixBlockForReuse(blocks, maxRow, maxCol, blen);
 
-							// copy submatrix to block
+							// copy sub matrix to block
 							src.slice(row_offset, row_offset + maxRow - 1, col_offset, col_offset + maxCol - 1, block);
 						}
 						else { // empty block (not on diagonal)