--- a/src/main/java/org/apache/sysds/runtime/io/WriterBinaryBlockParallel.java
+++ b/src/main/java/org/apache/sysds/runtime/io/WriterBinaryBlockParallel.java
@@ -21,22 +21,18 @@ package org.apache.sysds.runtime.io;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
 
 import org.apache.hadoop.fs.FileSystem;
-import org.apache.hadoop.fs.LocalFileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.mapred.JobConf;
-import org.apache.sysds.conf.DMLConfig;
 import org.apache.sysds.hops.OptimizerUtils;
 import org.apache.sysds.runtime.DMLRuntimeException;
 import org.apache.sysds.runtime.controlprogram.parfor.stat.InfrastructureAnalyzer;
 import org.apache.sysds.runtime.matrix.data.MatrixBlock;
 import org.apache.sysds.runtime.util.CommonThreadPool;
-import org.apache.sysds.runtime.util.HDFSTool;
 
 public class WriterBinaryBlockParallel extends WriterBinaryBlock
 {
@@ -45,71 +41,68 @@ public class WriterBinaryBlockParallel extends WriterBinaryBlock
 	}
 	
 	@Override
-	protected void writeBinaryBlockMatrixToHDFS( Path path, JobConf job, FileSystem fs, MatrixBlock src, long rlen, long clen, int blen )
+	protected void writeBinaryBlockMatrixToHDFS( Path path, JobConf job, MatrixBlock src, long rlen, long clen, int blen )
 		throws IOException, DMLRuntimeException
 	{
 		//estimate output size and number of output blocks (min 1)
-		int numPartFiles = (int)(OptimizerUtils.estimatePartitionedSizeExactSparsity(rlen, clen, 
-				blen, src.getNonZeros()) / InfrastructureAnalyzer.getHDFSBlockSize());
-		numPartFiles = Math.max(numPartFiles, 1);
-		
+		int numPartFiles = numPartsFiles(path.getFileSystem(job), rlen, clen, blen , src.getNonZeros());
+
 		//determine degree of parallelism
 		int numThreads = OptimizerUtils.getParallelBinaryWriteParallelism();
 		numThreads = Math.min(numThreads, numPartFiles);
 		
 		//fall back to sequential write if dop is 1 (e.g., <128MB) in order to create single file
 		if( numThreads <= 1 ) {
-			super.writeBinaryBlockMatrixToHDFS(path, job, fs, src, rlen, clen, blen);
+			super.writeBinaryBlockMatrixToHDFS(path, job,  src, rlen, clen, blen);
 			return;
 		}
 
 		//create directory for concurrent tasks
-		HDFSTool.createDirIfNotExistOnHDFS(path, DMLConfig.DEFAULT_SHARED_DIR_PERMISSION);
+		// HDFSTool.createDirIfNotExistOnHDFS(path, DMLConfig.DEFAULT_SHARED_DIR_PERMISSION);
 		
 		//create and execute write tasks
-		try 
-		{
-			ExecutorService pool = CommonThreadPool.get(numThreads);
+		final ExecutorService pool = CommonThreadPool.get(numThreads);
+		try {
 			ArrayList<WriteFileTask> tasks = new ArrayList<>();
 			int blklen = (int)Math.ceil((double)rlen / blen / numThreads) * blen;
 			for(int i=0; i<numThreads & i*blklen<rlen; i++) {
 				Path newPath = new Path(path, IOUtilFunctions.getPartFileName(i));
-				tasks.add(new WriteFileTask(newPath, job, fs, src, i*blklen, Math.min((i+1)*blklen, rlen), blen));
+				tasks.add(new WriteFileTask(newPath, job,  src, i*blklen, Math.min((i+1)*blklen, rlen), blen));
 			}
 
-			//wait until all tasks have been executed
-			List<Future<Object>> rt = pool.invokeAll(tasks);	
-			pool.shutdown();
-			
 			//check for exceptions 
-			for( Future<Object> task : rt )
+			for( Future<Object> task : pool.invokeAll(tasks) )
 				task.get();
 			
-			// delete crc files if written to local file system
-			if (fs instanceof LocalFileSystem) {
-				for(int i=0; i<numThreads & i*blklen<rlen; i++) 
-					IOUtilFunctions.deleteCrcFilesFromLocalFileSystem(fs,
-						new Path(path, IOUtilFunctions.getPartFileName(i)));
-			}
 		} 
 		catch (Exception e) {
 			throw new IOException("Failed parallel write of binary block input.", e);
 		}
+		finally{
+			pool.shutdown();
+		}
+	}
+
+	public static int numPartsFiles(FileSystem fs, long rlen, long clen, long blen, long nZeros) {
+		int numPartFiles = (int) (OptimizerUtils.estimatePartitionedSizeExactSparsity(rlen, clen, blen, nZeros) /
+			InfrastructureAnalyzer.getBlockSize(fs));
+		numPartFiles = Math.max(numPartFiles, 1);
+		numPartFiles = Math.min(numPartFiles,
+			(int) (Math.ceil((double) rlen / blen) * Math.ceil((double) clen / blen)));
+		return numPartFiles;
 	}
 
 	private class WriteFileTask implements Callable<Object> 
 	{
 		private Path _path = null;
 		private JobConf _job = null;
-		private FileSystem _fs = null;
 		private MatrixBlock _src = null;
 		private long _rl = -1;
 		private long _ru = -1;
 		private int _blen = -1;
 		
-		public WriteFileTask(Path path, JobConf job, FileSystem fs, MatrixBlock src, long rl, long ru, int blen) {
+		public WriteFileTask(Path path, JobConf job, MatrixBlock src, long rl, long ru, int blen) {
 			_path = path;
-			_fs = fs;
 			_job = job;
 			_src = src;
 			_rl = rl;
@@ -118,10 +111,9 @@ public class WriterBinaryBlockParallel extends WriterBinaryBlock
 		}
 	
 		@Override
-		public Object call() 
-			throws Exception 
-		{
-			writeBinaryBlockMatrixToSequenceFile(_path, _job, _fs, _src, _blen, (int)_rl, (int)_ru);
+		public Object call() throws Exception {
+			writeBinaryBlockMatrixToSequenceFile(_path, _job,  _src, _blen, (int) _rl, (int) _ru);
+			IOUtilFunctions.deleteCrcFilesFromLocalFileSystem(_job, _path);
 			return null;
 		}
 	}