--- a/src/main/java/org/apache/sysds/runtime/transform/encode/ColumnEncoder.java
+++ b/src/main/java/org/apache/sysds/runtime/transform/encode/ColumnEncoder.java
@@ -29,8 +29,8 @@ import java.io.ObjectOutput;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.Callable;
@@ -40,10 +40,11 @@ import org.apache.commons.logging.LogFactory;
 import org.apache.sysds.api.DMLScript;
 import org.apache.sysds.runtime.DMLRuntimeException;
 import org.apache.sysds.runtime.controlprogram.caching.CacheBlock;
-import org.apache.sysds.runtime.data.SparseRowVector;
-import org.apache.sysds.runtime.frame.data.FrameBlock;
+import org.apache.sysds.runtime.data.DenseBlock;
 import org.apache.sysds.runtime.data.SparseBlock;
 import org.apache.sysds.runtime.data.SparseBlockCSR;
+import org.apache.sysds.runtime.data.SparseRowVector;
+import org.apache.sysds.runtime.frame.data.FrameBlock;
 import org.apache.sysds.runtime.matrix.data.MatrixBlock;
 import org.apache.sysds.runtime.util.DependencyTask;
 import org.apache.sysds.runtime.util.DependencyThreadPool;
@@ -129,25 +130,24 @@ public abstract class ColumnEncoder implements Encoder, Comparable<ColumnEncoder
 
 	protected abstract double getCode(CacheBlock<?>in, int row);
 
-	protected abstract double[] getCodeCol(CacheBlock<?> in, int startInd, int blkSize);
-
-
-	/*protected void applySparse(CacheBlock in, MatrixBlock out, int outputCol, int rowStart, int blk){
-		int index = _colID - 1;
-		for(int r = rowStart; r < getEndIndex(in.getNumRows(), rowStart, blk); r++) {
-			SparseRowVector row = (SparseRowVector) out.getSparseBlock().get(r);
-			row.values()[index] = getCode(in, r);
-			row.indexes()[index] = outputCol;
-		}
-	}*/
+	/**
+	 * Get the coded values for a given range from start to end.
+	 * 
+	 * @param in       The CacheBlock to extract the values from
+	 * @param startInd The start Index
+	 * @param rowEnd   The end index
+	 * @param tmp 	   double tmp array to put the result into if valid.
+	 * @return The encoded double values.
+	 */
+	protected abstract double[] getCodeCol(CacheBlock<?> in, int startInd, int rowEnd, double[] tmp);
 
 	protected void applySparse(CacheBlock<?> in, MatrixBlock out, int outputCol, int rowStart, int blk){
 		boolean mcsr = MatrixBlock.DEFAULT_SPARSEBLOCK == SparseBlock.Type.MCSR;
 		mcsr = false; //force CSR for transformencode
 		int index = _colID - 1;
 		// Apply loop tiling to exploit CPU caches
-		double[] codes = getCodeCol(in, rowStart, blk);
 		int rowEnd = getEndIndex(in.getNumRows(), rowStart, blk);
+		double[] codes = getCodeCol(in, rowStart, rowEnd, null);
 		int B = 32; //tile size
 		for(int i = rowStart; i < rowEnd; i+=B) {
 			int lim = Math.min(i+B, rowEnd);
@@ -168,23 +168,37 @@ public abstract class ColumnEncoder implements Encoder, Comparable<ColumnEncoder
 		}
 	}
 
-	/*protected void applyDense(CacheBlock in, MatrixBlock out, int outputCol, int rowStart, int blk){
-		for(int i = rowStart; i < getEndIndex(in.getNumRows(), rowStart, blk); i++) {
-			out.quickSetValue(i, outputCol, getCode(in, i));
-		}
-	}*/
-	
 	protected void applyDense(CacheBlock<?> in, MatrixBlock out, int outputCol, int rowStart, int blk){
 		// Apply loop tiling to exploit CPU caches
-		double[] codes = getCodeCol(in, rowStart, blk);
-		int rowEnd = getEndIndex(in.getNumRows(), rowStart, blk);
-		int B = 32; //tile size
-		for(int i = rowStart; i < rowEnd; i+=B) {
-			int lim = Math.min(i+B, rowEnd);
-			for (int ii=i; ii<lim; ii++)
-				out.quickSetValue(ii, outputCol, codes[ii-rowStart]);
-				//out.denseSuperQuickSetValue(ii, outputCol, codes[ii-rowStart]);
+		final int rowEnd = getEndIndex(in.getNumRows(), rowStart, blk);
+		final int smallTile = 64;
+		final double[] tmp = new double[smallTile]; 
+		final DenseBlock outB = out.getDenseBlock();
+		if( outB.isContiguous(rowStart, blk)) 
+			for(int i = rowStart; i < rowEnd; i += smallTile) 
+				applyDenseTileContiguous(in, outB, outputCol, i, Math.min(i + smallTile, rowEnd),tmp);	
+		else 
+			for(int i = rowStart; i < rowEnd; i += smallTile) 
+				applyDenseTileGeneric(in, outB, outputCol, i, Math.min(i + smallTile, rowEnd),tmp);
+		
+	}
+
+	private void applyDenseTileContiguous(CacheBlock<?> in, DenseBlock out, int outputCol, int s, int e, double[] tmp) {
+		// these are codes for this block offset by rowStart
+		final double[] codes = getCodeCol(in, s, e, tmp);
+		final double[] vals = out.values(s);
+		int off = out.pos(s) + outputCol;
+		final int nCol = out.getDim(1);
+		for(int i = 0; i < e - s; i++, off += nCol){
+			vals[off] = codes[i];
 		}
+	}
+
+	private void applyDenseTileGeneric(CacheBlock<?> in, DenseBlock out, int outputCol, int s, int e, double[] tmp) {
+		// these are codes for this block offset by rowStart
+		final double[] codes = getCodeCol(in, s, e, tmp);
+		for(int i = s; i < e; i++) // rows
+			out.set(i, outputCol, codes[i - s]);
 	}
 
 	protected abstract TransformType getTransformType();