… src/main/java/org/apache/sysds/runtime/transform/encode/ColumnEncoderBin.java
… src/main/java/org/apache/sysds/runtime/transform/encode/ColumnEncoderDummycode.java
… src/main/java/org/apache/sysds/runtime/transform/encode/ColumnEncoderFeatureHash.java
… src/main/java/org/apache/sysds/runtime/transform/encode/ColumnEncoderPassThrough.java
… src/main/java/org/apache/sysds/runtime/transform/encode/ColumnEncoderUDF.java
… src/main/java/org/apache/sysds/runtime/transform/encode/ColumnEncoderWordEmbedding.java
… src/main/java/org/apache/sysds/runtime/transform/encode/CompressedEncode.java
… src/main/java/org/apache/sysds/runtime/transform/encode/EncoderFactory.java
… src/main/java/org/apache/sysds/runtime/transform/encode/LegacyEncoder.java
… src/main/java/org/apache/sysds/runtime/transform/encode/MultiColumnEncoder.java
… src/main/java/org/apache/sysds/runtime/util/DataConverter.java
… src/main/java/org/apache/sysds/runtime/util/UtilFunctions.java
… src/test/java/org/apache/sysds/performance/Main.java
… src/test/java/org/apache/sysds/performance/compression/TransformPerf.java
… src/test/java/org/apache/sysds/performance/generators/FrameTransformFile.java
… src/test/java/org/apache/sysds/test/component/compress/colgroup/scheme/CLAConstSchemeTest.java
… src/test/java/org/apache/sysds/test/component/compress/colgroup/scheme/CLADDCSchemeTest.java
… src/test/java/org/apache/sysds/test/component/compress/colgroup/scheme/CLAEmptySchemeTest.java
… src/test/java/org/apache/sysds/test/component/federated/FederatedTestUtils.java
… src/test/java/org/apache/sysds/test/component/frame/transform/TransformCompressedTestSingleColBinSpecific.java
… src/test/java/org/apache/sysds/test/component/misc/IOUtilFunctionsTest.java
… src/test/java/org/apache/sysds/test/component/misc/NumPartsFilesTest.java
… src/test/java/org/apache/sysds/test/functions/federated/primitives/FederatedRCBindTest.java